// src/app/model/fornecedor.ts
export class Fornecedor {

    identificador: string;
	cnpj: string ;
	endereco :string ;
	email :string ;
	cep : string ;
	numero :string ;
	cidade: string;
	bairro: string;
	estado: string;
	complemento :string;	
	nomeFornecedor :string;	
	telefone: string;
	celular: string;
	whatsapp: string;	
}